+++ ergotree-ir/src/mir/sigma_and.rs
@@ -1,4 +1,5 @@
-use super::expr::Expr;
+//! AND conjunction for sigma propositions
+
 use crate::serialization::op_code::OpCode;
 use crate::serialization::sigma_byte_reader::SigmaByteRead;
 use crate::serialization::sigma_byte_writer::SigmaByteWrite;
@@ -6,19 +7,21 @@ use crate::serialization::SerializationError;
 use crate::serialization::SigmaSerializable;
 use crate::types::stype::SType;
 
-/// Logical OR op on collection of SBoolean values
+use super::expr::Expr;
+
+/// AND conjunction for sigma propositions
 #[derive(PartialEq, Eq, Debug, Clone)]
-pub struct Or {
-    /// Input collection
-    pub input: Box<Expr>,
+pub struct SigmaAnd {
+    /// Collection of SSigmaProp
+    pub items: Vec<Expr>,
 }
 
-impl Or {
-    pub(crate) const OP_CODE: OpCode = OpCode::OR;
+impl SigmaAnd {
+    pub(crate) const OP_CODE: OpCode = OpCode::SIGMA_AND;
 
     /// Type
     pub fn tpe(&self) -> SType {
-        SType::SBoolean
+        SType::SSigmaProp
     }
 
     pub(crate) fn op_code(&self) -> OpCode {
@@ -26,40 +29,36 @@ impl Or {
     }
 }
 
-impl SigmaSerializable for Or {
+impl SigmaSerializable for SigmaAnd {
     fn sigma_serialize<W: SigmaByteWrite>(&self, w: &mut W) -> Result<(), std::io::Error> {
-        self.input.sigma_serialize(w)
+        self.items.sigma_serialize(w)
     }
 
     fn sigma_parse<R: SigmaByteRead>(r: &mut R) -> Result<Self, SerializationError> {
         Ok(Self {
-            input: Expr::sigma_parse(r)?.into(),
+            items: Vec::<Expr>::sigma_parse(r)?,
         })
     }
 }
 
-#[cfg(feature = "arbitrary")]
 /// Arbitrary impl
+#[cfg(feature = "arbitrary")]
 mod arbitrary {
-    use crate::mir::expr::arbitrary::ArbExprParams;
-
     use super::*;
-
+    use crate::mir::constant::Constant;
+    use proptest::collection::vec;
     use proptest::prelude::*;
 
-    impl Arbitrary for Or {
+    impl Arbitrary for SigmaAnd {
         type Strategy = BoxedStrategy<Self>;
-        type Parameters = usize;
+        type Parameters = ();
 
-        fn arbitrary_with(args: Self::Parameters) -> Self::Strategy {
-            any_with::<Expr>(ArbExprParams {
-                tpe: SType::SColl(SType::SBoolean.into()),
-                depth: args,
-            })
-            .prop_map(|input| Self {
-                input: input.into(),
-            })
-            .boxed()
+        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
+            vec(any_with::<Constant>(SType::SSigmaProp.into()), 2..5)
+                .prop_map(|constants| Self {
+                    items: constants.into_iter().map(|c| c.into()).collect(),
+                })
+                .boxed()
         }
     }
 }
@@ -74,7 +73,7 @@ mod tests {
     proptest! {
 
         #[test]
-        fn ser_roundtrip(v in any_with::<Or>(1)) {
+        fn ser_roundtrip(v in any::<SigmaAnd>()) {
             let expr: Expr = v.into();
             prop_assert_eq![sigma_serialize_roundtrip(&expr), expr];
         }