+++ ergotree-ir/src/serialization/serializable.rs
@@ -2,6 +2,7 @@
 use crate::mir::expr::InvalidArgumentError;
 use crate::mir::val_def::ValId;
 use crate::types::stuple::STupleItemsOutOfBoundsError;
+use crate::types::type_unify::TypeUnificationError;
 
 use super::{
     constant_store::ConstantStore,
@@ -81,6 +82,12 @@ impl From<InvalidArgumentError> for SerializationError {
 impl From<STupleItemsOutOfBoundsError> for SerializationError {
     fn from(e: STupleItemsOutOfBoundsError) -> Self {
         SerializationError::ValueOutOfBounds(format!("{:?}", e))
+    }
+}
+
+impl From<TypeUnificationError> for SerializationError {
+    fn from(e: TypeUnificationError) -> Self {
+        SerializationError::Misc(format!("{:?}", e))
     }
 }